+++ src/lib/inject.ts
@@ -17,7 +17,6 @@ const moment = require('moment-timezone');
 
 class Injector {
   public enabled: boolean;
-  public isDesktop: boolean;
   public notifyId: string;
   private spoof = {
     custom: '',
@@ -25,7 +24,7 @@ class Injector {
     metadata: {},
   };
 
-  constructor(settings: any, tempStore: any, profileCache: any, seed: number, isDesktop: boolean) {
+  constructor(settings: any, tempStore: any, profileCache: any, seed: number) {
     if (!settings.config.enabled) {
       this.enabled = false;
       return;
@@ -33,7 +32,6 @@ class Injector {
 
     this.enabled = true;
     this.notifyId = tempStore.notifyId;
-    this.isDesktop = isDesktop;
 
     // profile to be used for injection
     let p: any = null;
@@ -214,6 +212,8 @@ class Injector {
     (function(){
         if (
           window.location.href.startsWith("https://www.google.com/recaptcha/api2") || 
+          window.location.href.startsWith("https://accounts.google.com/")          ||
+          window.location.href.startsWith("https://accounts.youtube.com/")         ||
           window.location.href.startsWith("https://disqus.com/embed/comments/")    ||
           window.CHAMELEON_SPOOF
         ) { 
@@ -223,15 +223,6 @@ class Injector {
         let CHAMELEON_SPOOF = new WeakMap();
         CHAMELEON_SPOOF.set(window, JSON.parse(\`${JSON.stringify(this.spoof.metadata)}\`));
         window.CHAMELEON_SPOOF = Symbol.for("CHAMELEON_SPOOF");
-        window.CHAMELEON_SPOOF_FP_DETECTED = {
-          fpPanel: {
-            audioContext: false,
-            clientRects: false,
-            date: false,
-            screen: false,
-            webSocket: false
-          }
-        };
 
         let injectionProperties = JSON.parse(\`${JSON.stringify(this.spoof.overwrite)}\`);
 
@@ -283,9 +274,7 @@ class Injector {
         });
         
         let iframeWindow = HTMLIFrameElement.prototype.__lookupGetter__('contentWindow');
-        let frameWindow = HTMLFrameElement.prototype.__lookupGetter__('contentWindow');
         let iframeDocument = HTMLIFrameElement.prototype.__lookupGetter__('contentDocument');
-        let frameDocument = HTMLFrameElement.prototype.__lookupGetter__('contentDocument');
 
         ${this.spoof.custom}
 
@@ -293,41 +282,6 @@ class Injector {
           contentWindow: {
             get: function() {
               let f = iframeWindow.apply(this);
-              try {
-                if (f) {
-                  Object.defineProperty(f, 'Date', {
-                    value: window.Date
-                  });
-  
-                  Object.defineProperty(f.Intl, 'DateTimeFormat', {
-                    value: window.Intl.DateTimeFormat
-                  });
-  
-                  Object.defineProperty(f, 'screen', {
-                    value: window.screen
-                  });
-  
-                  Object.defineProperty(f, 'navigator', {
-                    value: window.navigator
-                  });
-                }
-              } catch (e) {}
-              
-              return f;
-            }
-          },
-          contentDocument: {
-            get: function() {
-              let f = iframeDocument.apply(this);
-              return f;
-            }
-          }
-        });
-
-        Object.defineProperties(HTMLFrameElement.prototype, {
-          contentWindow: {
-            get: function() {
-              let f = frameWindow.apply(this);
               if (f) {
                 try {
                   Object.defineProperty(f, 'Date', {
@@ -345,6 +299,22 @@ class Injector {
                   Object.defineProperty(f, 'navigator', {
                     value: window.navigator
                   });
+
+                  Object.defineProperty(f.Element.prototype, 'getBoundingClientRect', {
+                    value: window.Element.prototype.getBoundingClientRect
+                  });
+
+                  Object.defineProperty(f.Element.prototype, 'getClientRects', {
+                    value: window.Element.prototype.getClientRects
+                  });
+
+                  Object.defineProperty(f.Range.prototype, 'getBoundingClientRect', {
+                    value: window.Range.prototype.getClientRects
+                  });
+
+                  Object.defineProperty(f.Range.prototype, 'getClientRects', {
+                    value: window.Range.prototype.getClientRects
+                  });
                 } catch (e) {}
               }
               return f;
@@ -352,169 +322,11 @@ class Injector {
           },
           contentDocument: {
             get: function() {
-              let f = frameDocument.apply(this);
-              return f;
+              this.contentWindow;
+              return iframeDocument.apply(this);
             }
           }
         });
-
-        // Proxy access to objects for fingerprint panel in popup
-        if (${this.isDesktop}) {
-          // check if audio context is enabled
-          if (window.AudioContext) {
-            window.AudioContext = new Proxy(window.AudioContext, {  
-              construct: function(target, args) {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.audioContext) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.audioContext = true;
-                  window.top.postMessage({fp: 'audioContext', id: "${this.notifyId}"}, "*");
-                }
-                return new target(...args);
-              }
-            });
-
-            window.OfflineAudioContext = new Proxy(window.OfflineAudioContext, {  
-              construct: function(target, args) {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.audioContext) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.audioContext = true;
-                  window.top.postMessage({fp: 'audioContext', id: "${this.notifyId}"}, "*");
-                }
-                return new target(...args);
-              }
-            });
-          }
-
-          window.WebSocket = new Proxy(window.WebSocket, {  
-            construct: function(target, args) {
-              if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.webSocket) {
-                window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.webSocket = true;
-                window.top.postMessage({fp: 'webSocket', id: "${this.notifyId}"}, "*");
-              }
-              return new target(...args);
-            }
-          });
-
-          window.Date = new Proxy(window.Date, {  
-            construct: function(target, args) {
-              if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.date) {
-                window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.date = true;
-                window.top.postMessage({fp: 'date', id: "${this.notifyId}"}, "*");
-              }
-              return new target(...args);
-            }
-          });
-
-          ((innerHeight, innerWidth, outerHeight, outerWidth, clientHeight, clientWidth) => {
-            Object.defineProperty(window, 'innerHeight', {
-              configurable: true,
-              get: () => {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return innerHeight;
-              }
-            });
-            
-            Object.defineProperty(window, 'innerWidth', {
-              configurable: true,
-              get: () => {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return innerWidth;
-              }
-            });
-
-            Object.defineProperty(window, 'outerHeight', {
-              configurable: true,
-              get: () => {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return outerHeight;
-              }
-            });
-            
-            Object.defineProperty(window, 'outerWidth', {
-              configurable: true,
-              get: () => {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return outerWidth;
-              }
-            });
-
-            Object.defineProperty(window.document.documentElement, 'clientHeight', {
-              configurable: true,
-              get: () => {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return clientHeight;
-              }
-            });
-            
-            Object.defineProperty(window.document.documentElement, 'clientWidth', {
-              configurable: true,
-              get: () => {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return clientWidth;
-              }
-            });
-
-            window.screen = new Proxy(window.screen, {  
-              get: function(obj, prop) {
-                if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen) {
-                  window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.screen = true;
-                  window.top.postMessage({fp: 'screen', id: "${this.notifyId}"}, "*");
-                }
-                return obj[prop];
-              }
-            });
-          })(window.innerHeight, window.innerWidth, window.outerHeight, window.outerWidth, window.document.documentElement.clientHeight, window.document.documentElement.clientWidth);
-
-          ((getClientRects, getBoundingClientRect, rgetClientRects, rgetBoundingClientRect) => {
-            window.Element.prototype.getClientRects = function(){
-              if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects) {
-                window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects = true;
-                window.top.postMessage({fp: 'clientRects', id: "${this.notifyId}"}, "*");
-              }
-              return getClientRects.apply(this);
-            }
-
-            window.Element.prototype.getBoundingClientRect = function(){
-              if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects) {
-                window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects = true;
-                window.top.postMessage({fp: 'clientRects', id: "${this.notifyId}"}, "*");
-              }
-              return getBoundingClientRect.apply(this);
-            }
-
-            window.Range.prototype.getClientRects = function(){
-              if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects) {
-                window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects = true;
-                window.top.postMessage({fp: 'clientRects', id: "${this.notifyId}"}, "*");
-              }
-              return rgetClientRects.apply(this);
-            }
-
-            window.Range.prototype.getBoundingClientRect = function(){
-              if (!window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects) {
-                window.CHAMELEON_SPOOF_FP_DETECTED.fpPanel.clientRects = true;
-                window.top.postMessage({fp: 'clientRects', id: "${this.notifyId}"}, "*");
-              }
-              return rgetBoundingClientRect.apply(this);
-            }
-          })(window.Element.prototype.getClientRects, window.Element.prototype.getBoundingClientRect, window.Range.prototype.getClientRects, window.Range.prototype.getBoundingClientRect);
-        }
     })()
     `.replace(/CHAMELEON_SPOOF/g, chameleonObjName);
   }
@@ -529,7 +341,7 @@ class Injector {
 }
 
 // @ts-ignore
-let chameleonInjector = new Injector(settings, tempStore, profileCache, seed, isDesktop);
+let chameleonInjector = new Injector(settings, tempStore, profileCache, seed);
 
 if (chameleonInjector.enabled) {
   chameleonInjector.injectIntoPage();